Base

Base

Base

Material and particle properties

Types of waves

Effective wavenumbers

Reflection coefficients

Material and particle properties

EffectiveWaves.Medium — Type.

A homogenious background material.

source

EffectiveWaves.Specie — Type.

A circular or spherical specie of particle with homogenious material properties.

source

EffectiveWaves.volume_fraction — Function.

Returns the volume fraction of the specie.

source

EffectiveWaves.p_speed — Function.

Returns pressure wave speed, when β is the adiabatic bulk modulus/

source

EffectiveWaves.t_vectors — Function.

A t_matrix in the form of a vector, because for now we only deal with diagonal T matrices.

source

EffectiveWaves.Zn — Function.

Returns a ratio which gives the scattering strength of a particle

source

EffectiveWaves.maximum_hankel_order — Function.

Calculate the largest needed order for the hankel series.

source

Types of waves

There are currently three main types used.

EffectiveWaves.EffectiveWave — Type.

A type for the effective wave ansatz.

source

EffectiveWaves.effective_waves — Function.

Calculates the effective wavenumbers and return Vector{EffectiveWave}.

source

EffectiveWaves.AverageWave — Type.

A type for the ensemble average scattering coefficients. Here they are discretised in terms of the depth x of the halfspace

source

EffectiveWaves.average_error — Function.

Approximates the error in AverageWave.amplitudes due to the mesh AverageWave.x.

source

EffectiveWaves.x_mesh — Function.

Returns x the mesh used to discretise the integral equations.

source

EffectiveWaves.MatchWave — Type.

A type for matched waves.

source

EffectiveWaves.match_error — Function.

Calculates the difference between the match of MatchWave.effectivewaves and MatchWave.averagewave. This can be used as a proxi for convergence.

source

EffectiveWaves.x_mesh_match — Function.

Returns (x,L), where x[L:end] is the mesh used to match with wave_effs.

source

Effective wavenumbers

EffectiveWaves.wavenumbers — Function.

Returns all the transmitted effective wavenumbers

source

EffectiveWaves.wavenumber_low_volfrac — Function.

Explicit formula for one effective wavenumber based on a low particle volume fraction expasion.

source

EffectiveWaves.effective_medium — Function.

the effective low frequency bulk modulus and density of a material filled with particles

source

Reflection coefficients

EffectiveWaves.reflection_coefficient — Function.

Calculates the reflection coefficient from each wave in waves and then sums the results.

source

The average reflection coefficient

source

EffectiveWaves.reflection_coefficient_low_volfrac — Function.

An explicit formula for the refleciton coefficient based on a low particle volume fraction.

source

EffectiveWaves.reflection_coefficient_halfspace — Function.

the reflection from a homogenious halfspace, which is also the low frequency reflection from a particulate material when using the effective_medium.

source

EffectiveWaves.transmission_angle — Function.

calculate effective transmission angle θeff. We restrict -pi/2 < Re θeff < pi/2

source